Extract slippage input into separate component
